revert to inital editorStateSynced on migration
