+++ src/pageEditor/store/editor/editorSelectors/editorModSelectors.test.ts
@@ -17,7 +17,7 @@
 
 import { type EditorRootState } from "@/pageEditor/store/editor/pageEditorTypes";
 import { type ModComponentsRootState } from "@/store/modComponents/modComponentTypes";
-import { initialState as editorInitialState } from "@/pageEditor/store/editor/editorSlice";
+import { initialState as editorInitialState } from "@/store/editorInitialState";
 import { type ActivatedModComponent } from "@/types/modComponentTypes";
 import { type ModComponentFormState } from "@/pageEditor/starterBricks/formStateTypes";
 import {
